BUILD-8328 revert package name change
