+++ .github/workflows/release.yml
@@ -2,18 +2,16 @@ name: Release
 
 # Trigger when publishing a new GitHub release
 on:
-  push:
   release:
     types:
       - published
-  workflow_dispatch:
 
 jobs:
   release:
     permissions:
       id-token: write
       contents: write
-    uses: SonarSource/gh-action_release/.github/workflows/main.yaml@test/smarini/BUILD-8328
+    uses: SonarSource/gh-action_release/.github/workflows/main.yaml@34d8b20d125bfd58d124e84b007d3a18e61c358a # 5.10.4
     with:
       publishToBinaries: true
       mavenCentralSync: true
