+++ gradle.properties
@@ -1,4 +1,4 @@
-group = io.github.tomverin
+group = org.sonarsource.dummy
 version=2.7-SNAPSHOT
 projectType=application
 org.gradle.caching=true
