+++ sonar-dummy-oss-plugin/build.gradle
@@ -12,7 +12,7 @@ jar {
     attributes(
       'Plugin-Key': 'dummy-oss',
       'Plugin-Version': project.version,
-      'Plugin-Class': 'io.github.tomverin.plugin.DummyPlugin',
+      'Plugin-Class': 'org.sonarsource.dummy.plugin.DummyPlugin',
       'Plugin-ChildFirstClassLoader': 'false',
       'Sonar-Version': project.version,
       'SonarLint-Supported': 'true',
@@ -24,7 +24,7 @@ jar {
 sonarqube {
   properties {
     property 'sonar.projectName', 'SonarSource :: dummy :: Plugin'
-    property 'sonar.moduleKey', 'io.github.tomverin:sonar-dummy-oss-plugin'
+    property 'sonar.moduleKey', 'org.sonarsource.dummy:sonar-dummy-oss-plugin'
   }
 }
 
@@ -48,12 +48,12 @@ publishing {
           }
         }
         scm {
-          url = 'https://github.com/github/tomverin/sonar-dummy-oss-plugin'
+          url = 'https://github.com/SonarSource/sonar-dummy-oss-plugin'
         }
         developers {
           developer {
-            id = 'tomverin'
-            name = 'Tom Verin'
+            id = 'release-engineers'
+            name = 'SonarSource Release Engineers'
           }
         }
       }
