+++ sonar-dummy-oss-plugin/src/main/java/org/sonarsource/dummy/plugin/DummyPlugin.java
@@ -16,5 +16,24 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-@javax.annotation.ParametersAreNonnullByDefault
-package io.github.tomverin.dummy.plugin;
+package org.sonarsource.dummy.plugin;
+
+import java.util.List;
+import org.sonar.api.SonarPlugin;
+import com.google.common.collect.Lists;
+
+public final class DummyPlugin extends SonarPlugin {
+
+  @SuppressWarnings({"rawtypes", "unchecked"})
+  @Override
+  public List getExtensions() {
+    return Lists.newArrayList();
+  }
+
+
+  public String sayHello() {
+    System.out.println("hello");
+    return "hello";
+  }
+
+}
